TASK: Extract ReferencesProcessor from PropertiesProcessor
