+++ Classes/Application/Command/NodeTemplateCommandController.php
@@ -6,6 +6,8 @@ namespace Flowpack\NodeTemplates\Application\Command;
 
 use Flowpack\NodeTemplates\Domain\ExceptionHandling\CaughtExceptions;
 use Flowpack\NodeTemplates\Domain\NodeCreation\NodeCreationService;
+use Flowpack\NodeTemplates\Domain\NodeCreation\PropertiesProcessor;
+use Flowpack\NodeTemplates\Domain\NodeCreation\ReferencesProcessor;
 use Flowpack\NodeTemplates\Domain\NodeTemplateDumper\NodeTemplateDumper;
 use Flowpack\NodeTemplates\Domain\TemplateConfiguration\TemplateConfigurationProcessor;
 use Neos\ContentRepository\Domain\Service\ContextFactoryInterface;
@@ -103,7 +105,10 @@ class NodeTemplateCommandController extends CommandController
                 $caughtExceptions
             );
 
-            $nodeCreation = new NodeCreationService($subgraph);
+            $nodeCreation = new NodeCreationService(
+                new PropertiesProcessor(),
+                new ReferencesProcessor($subgraph)
+            );
             $nodeCreation->createMutatorCollection($template, $nodeType, $caughtExceptions);
 
             if ($caughtExceptions->hasExceptions()) {
