+++ Classes/Domain/NodeCreation/PropertiesProcessor.php
@@ -4,44 +4,17 @@ namespace Flowpack\NodeTemplates\Domain\NodeCreation;
 
 use Flowpack\NodeTemplates\Domain\ExceptionHandling\CaughtException;
 use Flowpack\NodeTemplates\Domain\ExceptionHandling\CaughtExceptions;
-use Neos\ContentRepository\Domain\Model\NodeInterface;
-use Neos\ContentRepository\Domain\Model\NodeType;
-use Neos\ContentRepository\Domain\Service\Context;
-use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Property\Exception as PropertyMappingException;
 use Neos\Flow\Property\PropertyMapper;
 use Neos\Flow\Property\PropertyMappingConfiguration;
 
-/**
- * @Flow\Proxy(false)
- */
 class PropertiesProcessor
 {
-    private Context $subgraph;
-
     private PropertyMapper $propertyMapper;
 
-    public function __construct(Context $subgraph, PropertyMapper $propertyMapper)
+    public function __construct(PropertyMapper $propertyMapper)
     {
-        $this->subgraph = $subgraph;
         $this->propertyMapper = $propertyMapper;
-    }
-
-    public function createFromArrayByTypeDeclaration(array $propertiesAndReferences, NodeType $nodeType): Properties
-    {
-        $references = [];
-        $properties = [];
-        foreach ($propertiesAndReferences as $propertyName => $propertyValue) {
-            // TODO: remove the next line to initialise the nodeType, once https://github.com/neos/neos-development-collection/issues/4333 is fixed
-            $nodeType->getFullConfiguration();
-            $declaration = $nodeType->getPropertyType($propertyName);
-            if ($declaration === 'reference' || $declaration === 'references') {
-                $references[$propertyName] = $propertyValue;
-                continue;
-            }
-            $properties[$propertyName] = $propertyValue;
-        }
-        return new Properties($properties, $references, $nodeType);
     }
 
     /**
@@ -101,57 +74,6 @@ class PropertiesProcessor
             }
         }
         return $validProperties;
-    }
-
-    public function processAndValidateReferences(TransientNode $node, CaughtExceptions $caughtExceptions): array
-    {
-        $nodeType = $node->getNodeType();
-        $validReferences = [];
-        foreach ($node->getReferences() as $referenceName => $referenceValue) {
-            $referenceType = ReferenceType::fromPropertyOfNodeType($referenceName, $nodeType);
-
-            try {
-                if ($referenceType->isReference()) {
-                    $nodeAggregateIdentifier = $referenceType->toNodeAggregateId($referenceValue);
-                    if ($nodeAggregateIdentifier === null) {
-                        continue;
-                    }
-                    if (!($resolvedNode = $this->subgraph->getNodeByIdentifier($nodeAggregateIdentifier->__toString())) instanceof NodeInterface) {
-                        throw new InvalidReferenceException(sprintf(
-                            'Node with identifier "%s" does not exist.',
-                            $nodeAggregateIdentifier->__toString()
-                        ), 1687632330292);
-                    }
-                    $validReferences[$referenceName] = $resolvedNode;
-                    continue;
-                }
-
-                if ($referenceType->isReferences()) {
-                    $nodeAggregateIdentifiers = $referenceType->toNodeAggregateIds($referenceValue);
-                    if (count($nodeAggregateIdentifiers) === 0) {
-                        continue;
-                    }
-                    $nodes = [];
-                    foreach ($nodeAggregateIdentifiers as $nodeAggregateIdentifier) {
-                        if (!($nodes[] = $this->subgraph->getNodeByIdentifier($nodeAggregateIdentifier->__toString())) instanceof NodeInterface) {
-                            throw new InvalidReferenceException(sprintf(
-                                'Node with identifier "%s" does not exist.',
-                                $nodeAggregateIdentifier->__toString()
-                            ), 1687632330292);
-                        }
-                    }
-                    $validReferences[$referenceName] = $nodes;
-                    continue;
-                }
-            } catch (InvalidReferenceException $runtimeException) {
-                $caughtExceptions->add(
-                    CaughtException::fromException($runtimeException)
-                        ->withOrigin(sprintf('Reference "%s" in NodeType "%s"', $referenceName, $nodeType->getName()))
-                );
-                continue;
-            }
-        }
-        return $validReferences;
     }
 
     /**
