+++ Classes/Domain/TemplateNodeCreationHandler.php
@@ -7,6 +7,8 @@ use Flowpack\NodeTemplates\Domain\ExceptionHandling\ExceptionHandler;
 use Flowpack\NodeTemplates\Domain\ExceptionHandling\TemplateNotCreatedException;
 use Flowpack\NodeTemplates\Domain\ExceptionHandling\TemplatePartiallyCreatedException;
 use Flowpack\NodeTemplates\Domain\NodeCreation\NodeCreationService;
+use Flowpack\NodeTemplates\Domain\NodeCreation\PropertiesProcessor;
+use Flowpack\NodeTemplates\Domain\NodeCreation\ReferencesProcessor;
 use Flowpack\NodeTemplates\Domain\TemplateConfiguration\TemplateConfigurationProcessor;
 use Neos\ContentRepository\Domain\Model\NodeInterface;
 use Neos\Flow\Annotations as Flow;
@@ -50,7 +52,11 @@ class TemplateNodeCreationHandler implements NodeCreationHandlerInterface
             $template = $this->templateConfigurationProcessor->processTemplateConfiguration($templateConfiguration, $evaluationContext, $caughtExceptions);
             $this->exceptionHandler->handleAfterTemplateConfigurationProcessing($caughtExceptions, $node);
 
-            $nodeMutators = (new NodeCreationService($node->getContext()))->createMutatorCollection($template, $node->getNodeType(), $caughtExceptions);
+            $nodeMutators = (new NodeCreationService(
+                new PropertiesProcessor(),
+                new ReferencesProcessor($node->getContext())
+            ))->createMutatorCollection($template, $node->getNodeType(), $caughtExceptions);
+
             $nodeMutators->apply($node);
             $this->exceptionHandler->handleAfterNodeCreation($caughtExceptions, $node);
         } catch (TemplateNotCreatedException|TemplatePartiallyCreatedException $templateCreationException) {
